Remove theme-util.js and refactor theme handling
The commit includes the removal of theme-util.js and the transition of theme handling to a CSS class-based approach, leveraging Tailwind CSS's darkMode feature. It also involves modifying theme switcher in ThemeSlider component and adding a script in the head of the BaseLayout to load the preferred theme from localStorage on initial load. Removal of several unnecessary old theme handling CSS rules from global.css file is also part of the update.